Изменить возможный размер файла на величину INT

+++ SoftwareManagement.Api.Domain/Models/ReleaseFile.cs
@@ -8,7 +8,7 @@ public class ReleaseFile
     public FileKind Kind { get; set; }
     public FileRuntimeVersion RuntimeVersion { get; set; }
     public string CheckSum { get; set; }
-    public ulong Size { get; set; }
+    public int Size { get; set; }
     public DateTime Uploaded { get; set; }
     public string Description { get; set; }
     public Guid ReleaseId { get; set; }
